feat: merge GetUserAccountInfoAction and EnsureUserAccountExistsAction to one Action
