+++ packages/common/src/actions/index.ts
@@ -1,5 +1,4 @@
 export * from "./transfer";
 export * from "./get-flow-price";
 export * from "./get-token-info";
-export * from "./get-user-account-info";
 export * from "./ensure-user-account-exists";
